fix unintended error shown in latest version of mypy
This chapter discusses using mypy to find static typing errors in code. However, when using mypy 0.67 on this code it also reports a logical error in the case where both optional parameters are None. Since this is somewhat confusing to new learners it seems sensible to just patch the code so it only displays the original intended error rather than these additional logic errors.